+++ Sources/Root/FilterRootStateController.swift
@@ -29,6 +29,7 @@ public class FilterRootStateController: UIViewController {
     private var selectionDataSource: FilterSelectionDataSource
     private let filterSelectionTitleProvider: FilterSelectionTitleProvider
     weak var delegate: FilterRootStateControllerDelegate?
+    var searchQuerySuggestionDataSource: SearchQuerySuggestionsDataSource?
 
     private lazy var loadingViewController = LoadingViewController(backgroundColor: .white, presentationDelay: 0)
 
@@ -36,6 +37,7 @@ public class FilterRootStateController: UIViewController {
 
     private lazy var filterRootViewController: FilterRootViewController = {
         let vc = FilterRootViewController(title: "", navigator: navigator, selectionDataSource: selectionDataSource, filterSelectionTitleProvider: filterSelectionTitleProvider, delegate: self)
+        vc.searchQuerySuggestionDataSource = searchQuerySuggestionDataSource
         return vc
     }()
 
@@ -88,6 +90,7 @@ public class FilterRootStateController: UIViewController {
         case let .loadFreshFilters(dataSource):
             filterRootViewController.remove()
             filterRootViewController = FilterRootViewController(title: "", navigator: navigator, selectionDataSource: selectionDataSource, filterSelectionTitleProvider: filterSelectionTitleProvider, delegate: self)
+            filterRootViewController.searchQuerySuggestionDataSource = searchQuerySuggestionDataSource
             currentFilterDataSource = dataSource
             state = .filter
         case let .newSelectionDataSource(newSelectionDataSource):
@@ -130,6 +133,22 @@ extension FilterRootStateController: FilterRootViewControllerDelegate {
 
     public func filterRootViewControllerShouldShowResults(_: FilterRootViewController) {
         delegate?.filterRootStateControllerShouldShowResults(self)
+    }
+}
+
+extension FilterRootStateController: AnyFilterViewController {
+    public var mainScrollableContentView: UIScrollView? {
+        guard state == .filter else {
+            return nil
+        }
+        return filterRootViewController.mainScrollableContentView
+    }
+
+    public var isMainScrollableViewScrolledToTop: Bool {
+        guard state == .filter else {
+            return true
+        }
+        return filterRootViewController.isMainScrollableViewScrolledToTop
     }
 }